+++ monocle/migrate/__init__.py
@@ -16,6 +16,9 @@
 
 from monocle.db.db import ELmonocleDB
 from monocle.db.db import dict_to_change_or_event
+from monocle import utils
+
+from typing import Dict, List
 
 
 class NotAvailableException(Exception):
@@ -50,10 +53,52 @@ def self_merge(elastic_conn, index) -> None:
                 to_update = []
 
 
+def missing_url_gerrit_events(elastic_conn, index) -> None:
+    bulk_size = 500
+    client = ELmonocleDB(elastic_conn, index)
+    url_cache: Dict[str, str] = {}
+
+    def get_change_url(obj: Dict) -> str:
+        change_id = obj["change_id"]
+        if change_id in url_cache:
+            return url_cache[change_id]
+        # Set a large size to avoid pagination, we do not expect more than 10000
+        # objects (change + events) for a given change
+        params = {"change_ids": [change_id], "size": 10000, "from": 0}
+        result = client.run_named_query("changes_and_events", ".*", params=params)
+        changes = list(filter(lambda obj: obj["type"] == "Change", result["items"]))
+        if len(changes) != 1:
+            raise RuntimeError("Wrong unicity for change %s" % change_id)
+        change = changes[0]
+        url = change["url"]
+        url_cache[change_id] = url
+        return url
+
+    def bulk_update(to_update: List) -> List:
+        print("Updating %s events ..." % len(to_update))
+        client.update(to_update)
+        return []
+
+    to_update = []
+    for _obj in client.iter_index():
+        obj = _obj["_source"]
+        if obj["type"] in utils.get_events_list() and "url" not in obj.keys():
+            url = get_change_url(obj)
+            obj["url"] = url
+            to_update.append(dict_to_change_or_event(obj))
+            if len(to_update) == bulk_size:
+                to_update = bulk_update(to_update)
+
+    bulk_update(to_update)
+
+
 def run_migrate(name, elastic_conn, index):
     if name not in processes:
         raise NotAvailableException()
     processes[name](elastic_conn, index)
 
 
-processes = {"self-merge": self_merge}
+processes = {
+    "self-merge": self_merge,
+    "missing-url-gerrit-events": missing_url_gerrit_events,
+}