+++ CHANGELOG.md
@@ -19,6 +19,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Upgrade
 
+- [readme] Added migration instructions for the missing url field in gerrit event objects.
+- [dbmanage] Added migration process: missing-url-gerrit-events
+
 ## [0.8.1] - 2021-02-15
 
 ### Fixed